requirements.txt
flask
tensorflow
tensorflow_hub
pillow
flask-cors
